trying to launch directly from local container. didn't work

+++ create-neo4j-host.sh
@@ -7,23 +7,32 @@ IMAGE=neo4j-server
 #USER=ariel.balter
 
 VM_NAME=$1
+VM_NAME=${VM_NAME:-balter-neo4j}
 
 #gcloud config set project ${PROJECT_ID}
 # gcloud config set zones ${zone}
 
-gcloud compute instances delete ${VM_NAME} \
-    --project ${PROJECT_ID} \
-    --zone ${ZONE} \
-    --delete-disks=boot
+# Delete instance if exists
+exsists=`gcloud compute instances list | grep "$VM_NAME"`
 
-# Create the vm in the psjh-eacri-biggraph project
-gcloud compute instances create ${VM_NAME} \
-    --image ${IMAGE} \
-    --image-project launcher-public \
-    --boot-disk-size=100GB \
+if [[ $exists != "" ]]; then
+	echo "exists: $exists......DELETING"
+	gcloud compute instances delete ${VM_NAME} \
+    	--project ${PROJECT_ID} \
+    	--zone ${ZONE} \
+    	--delete-disks=boot
+fi
+
+# Create the vm using the docker image
+cmd="
+gcloud compute instances create_with_container ${VM_NAME} \
+    --container-image ${IMAGE} \
+    --boot-disk-size=200GB \
     --boot-disk-type=pd-ssd \
     --machine-type=e2-highmem-2  \
-#    --network-interface=no-address \
     --zone=${ZONE} \
     --project=${PROJECT_ID}
+"
 
+echo "$cmd"
+$cmd
